Reorder socket options according to ZMQ header
No new options are added here, the existing ones are just shuffled
around so they match the order in which they're defined in the ZMQ C
header file. This is to make it easier to stay abreast of changes in
ZMQ, keeping track when options are added, deprecated or removed there.